refactor: build/order/create

+++ source/api/v1/services/order.service.js
@@ -2,7 +2,7 @@ import cron from 'node-cron'
 import Stripe from 'stripe'
 const stripe = new Stripe(`${process.env.STRIPE_SK}`)
 
-import { createOrder, detailOrderUUID, deleteOrderByUUID } from "../repositories/order.repo.js"
+import { createOrder, detailOrderUUID, deleteOrderByUUID, getOrderByAppointmentID } from "../repositories/order.repo.js"
 
 import { detailAppointmentUUID, updateAppointmentStatus } from "../repositories/appointment.repo.js"
 
@@ -68,9 +68,19 @@ export const create = async (req) => {
         }
     }
 
+    const existOrder = await getOrderByAppointmentID(appointment.id)
+    if (existOrder) {
+        const answer = {
+            status: 200,
+            info: {
+                msg: "Đơn hàng đã tồn tại. Vui lòng thanh toán!",
+            }
+        }
+        return answer
+    }
     const temp = await createOrder(newOrder)
     const order = await detailOrderUUID(temp.uuid)
-    
+
     const task = cron.schedule('*/15 * * * *', async () => {
         if (appointment.status_id == 1) {
             await deleteOrderByUUID(order.uuid)
